// src/types/Reciept.ts
export type Reciept = {
    account_number:            null | string;
    cashback:                  null;
    category:                  string;
    created_date:              Date;
    currency_code:             string;
    date:                      Date | null;
    delivery_date:             null;
    discount:                  number | null;
    document_reference_number: null | string;
    document_title:            null;
    document_type:             string;
    due_date:                  null;
    duplicate_of:              null;
    external_id:               null;
    id:                        number;
    img_file_name:             string;
    img_thumbnail_url:         string;
    img_url:                   string;
    insurance:                 null;
    invoice_number:            null | string;
    is_duplicate:              boolean;
    is_money_in:               boolean;
    line_items:                LineItem[];
    meta:                      Meta;
    notes:                     null;
    ocr_text:                  string;
    order_date:                null;
    payment:                   Payment;
    pdf_url:                   string;
    purchase_order_number:     null;
    reference_number:          string;
    rounding:                  null;
    service_end_date:          null;
    service_start_date:        null;
    shipping:                  null;
    store_number:              null | string;
    subtotal:                  number;
    tags:                      any[];
    tax:                       number | null;
    tax_lines:                 TaxLine[];
    tip:                       null;
    total:                     number;
    total_weight:              null;
    tracking_number:           null;
    updated_date:              Date;
    vendor:                    Vendor;
}

export type LineItem = {
    date:                   null;
    description:            null | string;
    discount:               number | null;
    discount_rate:          null;
    end_date:               null;
    full_description:       null | string;
    hsn:                    null;
    id:                     number;
    lot:                    null;
    normalized_description: null | string;
    order:                  number;
    price:                  number | null;
    quantity:               number;
    reference:              null;
    section:                null | string;
    sku:                    null | string;
    start_date:             null;
    tags:                   any[];
    tax:                    number | null;
    tax_rate:               null;
    text:                   string;
    total:                  number | null;
    type:                   Type | null;
    unit_of_measure:        null | string;
    upc:                    null;
    weight:                 null;
}

export enum Type {
    Alcohol = "alcohol",
    Discount = "discount",
    Fee = "fee",
    Food = "food",
    Product = "product",
}

export type Meta = {
    language:         string[];
    ocr_score:        number;
    owner:            string;
    pages:            Array<null[] | PageClass>;
    processed_pages:  number;
    source:           string;
    source_documents: Array<null[] | SourceDocumentClass>;
    total_pages:      number;
}

export type PageClass = {
    height:   number;
    language: string[];
    width:    number;
}

export type SourceDocumentClass = {
    height:  number;
    size_kb: number;
    width:   number;
}

export type Payment = {
    card_number:  null | string;
    display_name: null | string;
    terms:        null;
    type:         null | string;
}

export type TaxLine = {
    base:  null;
    name:  null;
    order: number;
    rate:  number;
    total: number;
}

export type Vendor = {
    account_number: null;
    address:        null | string;
    category:       null | string;
    email:          null | string;
    lat:            number | null;
    lng:            number | null;
    logo:           string;
    name:           string;
    phone_number:   null | string;
    raw_address:    null | string;
    raw_name:       string;
    type:           null | string;
    vat_number:     null | string;
    web:            null | string;
}